feat: plugin data flow e2e test case draft
