+++ plugins/gitlab/e2e/rawdata/_raw_gitlab_api_project.csv
@@ -1,0 +1,2 @@
+"id","params","data","url","input","created_at"
+2,"{""ProjectId"":3472737}","{""id"": 3472737, ""name"": ""inkscape"", ""path"": ""inkscape"", ""_links"": {""self"": ""https://gitlab.com/api/v4/projects/3472737"", ""events"": ""https://gitlab.com/api/v4/projects/3472737/events"", ""issues"": ""https://gitlab.com/api/v4/projects/3472737/issues"", ""labels"": ""https://gitlab.com/api/v4/projects/3472737/labels"", ""members"": ""https://gitlab.com/api/v4/projects/3472737/members"", ""repo_branches"": ""https://gitlab.com/api/v4/projects/3472737/repository/branches"", ""cluster_agents"": ""https://gitlab.com/api/v4/projects/3472737/cluster_agents"", ""merge_requests"": ""https://gitlab.com/api/v4/projects/3472737/merge_requests""}, ""mirror"": false, ""topics"": [], ""web_url"": ""https://gitlab.com/inkscape/inkscape"", ""archived"": false, ""tag_list"": [], ""namespace"": {""id"": 470642, ""kind"": ""group"", ""name"": ""Inkscape"", ""path"": ""inkscape"", ""web_url"": ""https://gitlab.com/groups/inkscape"", ""full_path"": ""inkscape"", ""parent_id"": null, ""avatar_url"": ""/uploads/-/system/group/avatar/470642/inkscape.png""}, ""avatar_url"": ""https://gitlab.com/uploads/-/system/project/avatar/3472737/inkscape.png"", ""created_at"": ""2017-06-09T14:16:35.615Z"", ""creator_id"": 402261, ""empty_repo"": false, ""readme_url"": ""https://gitlab.com/inkscape/inkscape/-/blob/master/README.md"", ""star_count"": 2627, ""visibility"": ""public"", ""description"": ""Inkscape vector image editor"", ""forks_count"": 966, ""lfs_enabled"": false, ""permissions"": {""group_access"": null, ""project_access"": null}, ""public_jobs"": true, ""jobs_enabled"": true, ""merge_method"": ""ff"", ""wiki_enabled"": false, ""build_timeout"": 18600, ""import_status"": ""none"", ""squash_option"": ""default_off"", ""ci_config_path"": null, ""default_branch"": ""master"", ""issues_enabled"": true, ""emails_disabled"": null, ""issues_template"": ""<!-- Please report new issues at https://inkscape.org/report; this tracker is for staff-confirmed issues only.\r\n     See our full bug reporting guidelines at https://inkscape.org/contribute/report-bugs/ -->\r\n\r\n#### Steps to reproduce:\r\n<!-- Describe what you did (step-by-step) so we can reproduce: -->\r\n\r\n- open Inkscape\r\n- ...\r\n\r\n#### What happened?\r\n\r\n...\r\n\r\n#### What should have happened?\r\n\r\n...\r\n\r\n#### Inkscape Version and Operating System:\r\n\r\n- Inkscape Version: ... <!-- (run inkscape -V or copy from Help → About Inkscape, top right) -->\r\n- Operating System: ...\r\n- Operating System version: ...\r\n\r\n<!-- Example file:\r\nAttach a sample file (or files) highlighting the issue, if appropriate. -->"", ""ssh_url_to_repo"": ""git@gitlab.com:inkscape/inkscape.git"", ""http_url_to_repo"": ""https://gitlab.com/inkscape/inkscape.git"", ""last_activity_at"": ""2022-05-05T08:26:20.527Z"", ""packages_enabled"": false, ""snippets_enabled"": false, ""open_issues_count"": 1472, ""wiki_access_level"": ""disabled"", ""pages_access_level"": ""enabled"", ""shared_with_groups"": [{""group_id"": 4575604, ""expires_at"": null, ""group_name"": ""Inkscape Developers"", ""group_full_path"": ""inkscape/devel"", ""group_access_level"": 30}, {""group_id"": 4445779, ""expires_at"": null, ""group_name"": ""Inkscape Bug Wranglers"", ""group_full_path"": ""inkscape/bug-wranglers"", ""group_access_level"": 20}], ""auto_devops_enabled"": false, ""builds_access_level"": ""enabled"", ""issues_access_level"": ""enabled"", ""name_with_namespace"": ""Inkscape / inkscape"", ""path_with_namespace"": ""inkscape/inkscape"", ""ci_default_git_depth"": null, ""forking_access_level"": ""enabled"", ""keep_latest_artifact"": true, ""merge_trains_enabled"": false, ""requirements_enabled"": true, ""service_desk_enabled"": null, ""compliance_frameworks"": [], ""merge_commit_template"": null, ""snippets_access_level"": ""disabled"", ""analytics_access_level"": ""enabled"", ""approvals_before_merge"": 0, ""marked_for_deletion_at"": null, ""marked_for_deletion_on"": null, ""merge_requests_enabled"": true, ""request_access_enabled"": false, ""shared_runners_enabled"": true, ""squash_commit_template"": null, ""merge_pipelines_enabled"": true, ""merge_requests_template"": """", ""operations_access_level"": ""disabled"", ""repository_access_level"": ""enabled"", ""requirements_access_level"": ""disabled"", ""suggestion_commit_message"": """", ""ci_job_token_scope_enabled"": false, ""container_registry_enabled"": false, ""auto_devops_deploy_strategy"": ""continuous"", ""autoclose_referenced_issues"": true, ""can_create_merge_request_in"": true, ""merge_requests_access_level"": ""enabled"", ""auto_cancel_pending_pipelines"": ""enabled"", ""ci_forward_deployment_enabled"": null, ""allow_merge_on_skipped_pipeline"": false, ""container_registry_access_level"": ""disabled"", ""container_registry_image_prefix"": ""registry.gitlab.com/inkscape/inkscape"", ""restrict_user_defined_variables"": false, ""security_and_compliance_enabled"": false, ""remove_source_branch_after_merge"": true, ""runner_token_expiration_interval"": null, ""resolve_outdated_diff_discussions"": false, ""printing_merge_request_link_enabled"": true, ""security_and_compliance_access_level"": ""disabled"", ""only_allow_merge_if_pipeline_succeeds"": false, ""external_authorization_classification_label"": """", ""only_allow_merge_if_all_discussions_are_resolved"": true}",https://gitlab.com/api/v4/projects/3472737?page=1&per_page=100&with_stats=true,"null","2022-05-05 09:38:23.165000000"
