+++ plugins/gitlab/e2e/rawdata/_tool_gitlab_projects.csv
@@ -1,0 +1,2 @@
+"gitlab_id","name","description","default_branch","path_with_namespace","web_url","creator_id","visibility","open_issues_count","star_count","forked_from_project_id","forked_from_project_web_url","created_date","updated_date","created_at","updated_at","_raw_data_params","_raw_data_table","_raw_data_id","_raw_data_remark"
+3472737,inkscape,Inkscape vector image editor,master,inkscape/inkscape,https://gitlab.com/inkscape/inkscape,0,public,1472,2627,0,"","2017-06-09 14:16:35.615000000","2022-05-05 08:26:20.527000000","2022-05-05 09:38:23.184000000","2022-05-05 09:38:23.184000000","{""ProjectId"":3472737}",_raw_gitlab_api_project,2,""
